fixed demo and made component for update form
