+++ frontend/src/components/Events/EachEvent.js
@@ -1,8 +1,9 @@
 import React, { useEffect, useState } from "react";
 import { useDispatch, useSelector } from "react-redux";
-import { useHistory, useParams, Route, NavLink } from "react-router-dom";
+import { useHistory, useParams } from "react-router-dom";
 import { getEventsThunk, deleteEventsThunk, updateEventsThunk } from "../../store/events";
 import './eachevent.css'
+import UpdateForm from "./updateform";
 
 export default function () {
     const dispatch = useDispatch()
@@ -22,39 +23,16 @@ export default function () {
     const event = events[eventId]
 
 
-    const [title, setTitle] = useState(event.title)
-    const [description, setDescription] = useState(event.description)
 
-
-    const [hidden, setHidden] = useState('hidden')
-    const [errors, setErrors] = useState([])
 
     useEffect(() => {
         dispatch(getEventsThunk())
         setEvend(event)
-    }, [dispatch])
-
-    let payload;
-    useEffect(() => {
-        const err = []
-        if (title.length <= 5) err.push('Title must be longer than 5 characters')
-        if (title.length >= 50) err.push('Title must be under 50 characters')
-        if (description.length <= 15) err.push('Description must be longer than 15 characters')
-        setErrors(err)
+    }, [dispatch, evend])
 
 
-    }, [title, description, event])
 
-    const onSubmit = (e) => {
-        e.preventDefault()
-        payload = {
-            id: event.id,
-            title,
-            description
-        }
-        dispatch(updateEventsThunk(payload))
-        setHidden('hidden')
-    }
+
 
     return (
         <>
@@ -66,28 +44,8 @@ export default function () {
                     history.push('/')
                 }}>Delete</button>
             }
-            <button onClick={() => {
-                if (hidden === 'hidden') {
-                    setHidden('nothidden')
-                } else {
-                    setHidden('hidden')
-                }
-            }
-            }>Edit</button>
-            <form onSubmit={onSubmit} className={hidden}>
-                {errors && errors.map(er => {
-                    return (
-                        <div>{er}</div>
-                    )
-                })}
-                <label>Title
-                    <input onChange={(e) => { setTitle(e.target.value) }} type='text' value={title}></input>
-                </label>
-                <label>Description
-                    <textarea onChange={(e) => { setDescription(e.target.value) }} type='text' value={description}></textarea>
-                </label>
-                <button>Submit Changes</button>
-            </form>
+
+            <UpdateForm event={event} />
             <h1>{event && event.title}</h1>
             <p>{event && event.description}</p>
         </>
